Reorganized LSTM code into parser and driver code
Also reorganized build files accordingly so that more targets can be
added, and made some globals member variables
