+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-project(cnn)
+project(lstm-parser)
 cmake_minimum_required(VERSION 2.8 FATAL_ERROR)
 
 set(CMAKE_MODULE_PATH ${PROJECT_SOURCE_DIR}/cmake)
@@ -26,6 +26,9 @@ include_directories(${EIGEN3_INCLUDE_DIR})
 #configure_file(${CMAKE_CURRENT_SOURCE_DIR}/config.h.cmake ${CMAKE_CURRENT_BINARY_DIR}/config.h)
 
 add_subdirectory(cnn/cnn)
-# add_subdirectory(cnn/examples)
-add_subdirectory(parser)
 
+set(COMMON_SOURCES parser/lstm-parse.cc)
+set(PARSER_SOURCES parser/lstm-parse-driver.cc)
+
+add_executable(lstm-parser ${COMMON_SOURCES} ${PARSER_SOURCES})
+target_link_libraries(lstm-parser ${LIBS} cnn)
